Moved core module back to parent. Kept examples module to allow build to pass. #6104
