+++ pulsar-function-go/conf/conf.go
@@ -26,7 +26,7 @@ import (
 	"os"
 	"time"
 
-	log "github.com/apache/pulsar/pulsar-function-go/core/logutil"
+	log "github.com/apache/pulsar/pulsar-function-go/logutil"
 	"gopkg.in/yaml.v2"
 )
