+++ pulsar-function-go/examples/outputFunc/outputFunc.go
@@ -22,7 +22,7 @@ package main
 import (
 	"context"
 
-	"github.com/apache/pulsar/pulsar-function-go/core/pf"
+	"github.com/apache/pulsar/pulsar-function-go/pf"
 )
 
 func HandleResponse(ctx context.Context, in []byte) ([]byte, error) {
