+++ pulsar-function-go/pb/generate.sh
